Adds TwitterApi and retrofit/retrofit-mock
Added part of the remote component to support the Twitter client
side of things.  Currently using retrofit-mock to fake network
request.  This is a first past, more work needs to be done here
maybe to clean things up.  I've added a TwitterApiPersistence which
I'm not using the moment.  Might come back to use it and simply remove
it.