Extracted interfaces for most data class
This allows the domain package to drive it's dependencies
as interface.  We then have the data package implement
those interface from the data package.  This simulates
the data package being a completely different application
the plugs into the domain app.

This is just an example of how to properly represent the domain
as the application and everything else simply plugs into the app.